+++ src/app/core/services/bookmark.service.ts
@@ -1,5 +1,5 @@
 import { inject, Injectable } from '@angular/core';
-import { map } from 'rxjs/operators';
+import { computed } from '@angular/core';
 
 import { ConfigService } from './config.service';
 
@@ -7,9 +7,5 @@ import { ConfigService } from './config.service';
 export class BookmarkService {
   private configService = inject(ConfigService);
 
-  readonly bookmarks$ = this.configService.config$.pipe(map((config) => config.bookmarks));
-
-  get bookmarks() {
-    return this.configService.subject.value?.bookmarks || [];
-  }
+  readonly bookmarks = computed(() => this.configService.config()?.bookmarks ?? []);
 }